django_project/requirements.txt
Django>=3.0,<4.0
Pillow>=8.0.0
django-crispy-forms>=1.9.0
crispy-bootstrap4>=1.0.0
django-allauth>=0.50.0
requests>=2.25.0
PyJWT>=2.0.0
cryptography>=3.0
